[#795] fix offset parameter in spring data rest integration

+++ integration/spring-data/base/src/main/java/com/blazebit/persistence/spring/data/repository/KeysetPageRequest.java
@@ -87,7 +87,7 @@ public class KeysetPageRequest extends PageRequest implements KeysetPageable {
 
     @Override
     public int getIntOffset() {
-        return offset;
+        return (int) offset;
     }
 
     @Override
